requirements.txt
lxml==4.9.3
requests==2.31.0
onepush==1.3.0
PyYAML==6.0.1